GCN design elements should not be inherited by labs route
